// stories/defaultStyles.d.ts
export declare const catColors: {
    background: string;
    accent: string;
};
export declare const defaultStyles: {
    title: {
        fontFamily: string;
        color: string;
    };
    textBlock: {
        width: string;
        fontSize: string;
        lineHeight: string;
    };
    link: {
        color: string;
    };
    accentText: {
        fontWeight: number;
        color: string;
    };
    accentHawkinsText: {
        fontWeight: number;
        color: string;
    };
};